requirements.txt
appdirs==1.4.3
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
cryptography==1.7.2
Django==1.11
django-cors-middleware==1.3.1
django-debug-toolbar==1.7
django-filter==1.0.2
django-rest-swagger==2.1.2
django-treebeard==4.1.2
djangorestframework==3.6.2
drfdocs==0.0.11
idna==2.5
itypes==1.1.0
Jinja2==2.9.6
jwt==0.5.2
Markdown==2.6.8
MarkupSafe==1.0
mysqlclient==1.3.10
openapi-codec==1.3.1
packaging==16.8
pyasn1==0.2.3
pycparser==2.17
Pygments==2.2.0
PyJWT==1.5.0
pyparsing==2.2.0
pytz==2017.2
requests==2.18.1
simplejson==3.11.1
six==1.10.0
sqlparse==0.2.3
uritemplate==3.0.0
urllib3==1.21.1
